Updated passage generation
Every cell in the dimension of the maze is now turned into a passage,
instead of leaving entire cells as walls. Cells now have four distinct
walls that determine from which sides they can be travelled to. Quick
generation boolean added to MazePanel to allow skipping of the
generation animation and printing of total execution time to the
console.